sanity check
